+++ src/client/components/results/class-result/class-result.html
@@ -25,8 +25,10 @@
                 :index="index"
                 :key="result.resultId"
                 :highlighted="index < highlightPlaces">
-            </result-row
-            >
+            </result-row>
         </tbody>
     </table>
+    <div v-if="onlyPlacings" class="has-text-centered">
+        <button class="button is-small" @click="showAll">Show All Placings</button>
+    </div>
 </div>
