+++ src/client/pages/home/home.html
@@ -12,6 +12,7 @@
                     <strong class="is-size-4">{{ show.name }}</strong>
                     <small>{{ formatDate(show.showDate) }}</small>
                 </a>
+                <hr>
             </li>
         </ul>
     </div>
