+++ src/server/api/result.ts
@@ -9,6 +9,7 @@ export function resultRouter(): express.Router {
     router.get("/byShow/:showId", resultController.getResultByShow);
     router.get("/byShow/:showId/:showClassId", resultController.getResultByShowClass);
     router.get("/placing/:showId/:showClassId", resultController.getPlacingByShowClass);
+    router.get("/placing/top/:showId/:showClassId", resultController.getTopPlacingByShowClass);
     router.post("/", resultController.createResult);
     router.put("/:id", resultController.updateResult);
     router.delete("/:id", resultController.deleteResult);
